+++ slm_lab/agent/memory/__init__.py
@@ -6,4 +6,4 @@ Contains different ways of storing an agents experiences and sampling from them
 # expose all the classes
 from .replay import *
 from .onpolicy import *
-from .prioritizedreplay import *
+from .prioritized import *
